Reapply: move `jax.interpreters.ad` to `jax._src.interpreters.ad`
Re-export roughly all of the same symbols via `jax.interpreters.ad` for now.

This version of the PR includes the names jax.interpreters.ad.source_info_util and jax.interpreters.ad.config, which the neural tangents is using.

PiperOrigin-RevId: 506642132
