+++ setup.cfg
@@ -35,6 +35,7 @@ per-file-ignores =
   jax/dtypes.py:F401
   jax/errors.py:F401
   jax/flatten_util.py:F401
+  jax/interpreters/ad.py:F401
   jax/linear_util.py:F401
   jax/prng.py:F401
   jax/profiler.py:F401
